+++ Changelog.md
@@ -7,6 +7,7 @@
 - Fix issue where alias text was getting trimmed before regex match
 - Add getregexmatch expression
 - Fix AutoTinker now allows gold to be added to Armor/Clothing
+- Fix bug where newly inscribed items were not seen by ub, allowing them to be sold/salvaged
 
 ## 0.2.2 (2021-01-16) [UtilityBeltInstaller-0.2.2.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/a597bed6c0330841d223ce2a762ca99d/UtilityBeltInstaller-0.2.2.exe)
 - Revert Fix `-` character must now be escaped in expression strings